+++ controllers/label_event_handler.go
@@ -1,4 +1,4 @@
-package controller
+package controllers
 
 import (
 	"k8s.io/apimachinery/pkg/types"
