+++ controllers/label_event_handler_test.go
@@ -1,4 +1,4 @@
-package controller
+package controllers
 
 import (
 	"testing"
